fix(whisper): fp32 on WebGPU ends the garbage transcriptions; hallucination guard; labelled logs
Root cause (provable regression from the June 'Update whisper impl.'
commit): dtype fp16 on the WebGPU load of whisper-tiny. Word-level
timestamps are computed by DTW over CROSS-ATTENTION tensors — text
survives fp16, the attention numerics do not, so every word time
collapses onto its segment end; _decode_asr then caps ends to the
segment timestamp token, producing the observed 20ms [t, t+0.02] spans,
and our midpoint dedupe sculpts that into repeated line-final tokens
('that.', 'this.', ' idea.') — one degenerate word per segment, 0
usable alignment. On Chrome the WebGPU path always wins, so the healthy
WASM q8 fallback never ran. dtype is fp32 now (upstream's own WebGPU
whisper examples pin fp32 for exactly this reason).

Defense in depth in the transcription hook:
- pure computeWhisperChunkPlan / sliceWhisperChunk (copies, not views;
  no more materializing all 44 chunks of a 66MB buffer upfront;
  redundant trailing windows skipped) / offsetWhisperSegments (a null
  end no longer coerces into an inverted segment) — all tested.
- hallucination detector (>60 percent identical normalized text across
  8+ segments, or median duration under 100ms): on detection nothing is
  saved or reported as success — poisoned IndexedDB caches are purged
  (incl. entries persisted by pre-guard runs), status goes to error
  with a user-facing message, now rendered in the Pitch Testing tab.
- all-chunks-failed errors instead of reporting done with 0 segments.
- optional label in every transcription log line — the Pitch Testing
  tab passes the file name, the mixer the song title (owner asked after
  chasing a garbage run with no song identity in the logs).

33 tests (chunk-plan invariants incl. the 18-minute/44-chunk case, both
detector rules at their thresholds, guard wiring incl. cache purge and
callback suppression).

+++ src/workers/whisper-worker.ts
@@ -72,7 +72,13 @@ async function loadModel() {
         'Xenova/whisper-tiny',
         {
           device: 'webgpu', // Try WebGPU first if supported
-          dtype: 'fp16',
+          // fp32, NOT fp16: word-level timestamps come from DTW over the
+          // cross-attention tensors, and fp16 corrupts those numerics —
+          // the text still decodes but every word time collapses onto
+          // the segment end (20ms spans, repeated line-final tokens; the
+          // owner's "that."/" idea." garbage). Upstream's own WebGPU
+          // whisper examples pin fp32 for exactly this reason.
+          dtype: 'fp32',
           progress_callback: handleProgressCallback,
         },
       )
